// src/app/core/interfaces/product-variation-review/product-variation-review.interface.ts
export interface ProductVariationReviewInterface {
    cmp_uuid: string;
    pro_uuid: string;
    prov_uuid: string;
    provrev_uuid: string;
    usr_uuid: string;
    cus_uuid: string;
    provrev_rating: number;
    provrev_comment: string;
    provrev_isverified: boolean;
    provrev_createdat: Date;
    provrev_updatedat: Date;
}